@@ -1,6 +1,6 @@
 source 'https://rubygems.org'
 
-gem 'rails', '~> 4.0.2'
+gem 'rails', '~> 4.1.0'
 gem 'sass-rails', '~> 4.0.0'
 gem 'coffee-rails', '~> 4.0.0'
 gem 'pg'
@@ -11,9 +11,9 @@ gem 'uglifier', '>= 1.0.3'
 gem 'zip-zip', '~> 0.2'
 
 group :development, :test do
-  gem 'rspec', '~> 2.13.0'
-  gem 'rspec-rails', '~> 2.13.1'
-  gem 'capybara', '~> 2.1.0'
-  gem 'selenium-webdriver', '~> 2.33.0'
-  gem 'rubocop', '~> 0.16.0'
+  gem 'rspec', '~> 3.0'
+  gem 'rspec-rails', '~> 3.0'
+  gem 'capybara', '~> 2.3.0'
+  gem 'selenium-webdriver', '~> 2.42.0'
+  gem 'rubocop', '~> 0.23.0'
 end
