@@ -4,8 +4,8 @@ describe Acme::Post do
   [true, false].each do |reticulated|
     it "POST #{reticulated ? 'reticulated' : 'unreticulated' } spline" do
       post '/api/spline', { 'reticulated' => reticulated }.to_json, 'Content-Type' => 'application/json'
-      response.status.should eq 201
-      response.body.should eq({ 'reticulated' => reticulated }.to_json)
+      expect(response.status).to eq 201
+      expect(response.body).to eq({ 'reticulated' => reticulated }.to_json)
     end
   end
 
