@@ -4,8 +4,8 @@ describe Acme::Protected do
   context 'without authorization' do
     it 'ping' do
       get '/api/protected/ping'
-      response.status.should eq 401
-      response.body.should eq ''
+      expect(response.status).to eq 401
+      expect(response.body).to eq ''
     end
   end
   context 'with incorrect authorization' do
@@ -13,8 +13,8 @@ describe Acme::Protected do
       get '/api/protected/ping', nil,
           'HTTP_AUTHORIZATION' => ActionController::HttpAuthentication::Basic.encode_credentials('foo', 'bar')
 
-      response.status.should eq 401
-      response.body.should eq ''
+      expect(response.status).to eq 401
+      expect(response.body).to eq ''
     end
   end
   context 'with authorization' do
@@ -22,8 +22,8 @@ describe Acme::Protected do
       get '/api/protected/ping', nil,
           'HTTP_AUTHORIZATION' => ActionController::HttpAuthentication::Basic.encode_credentials('username', 'password')
 
-      response.status.should eq 200
-      response.body.should eq({ ping: 'pong' }.to_json)
+      expect(response.status).to eq 200
+      expect(response.body).to eq({ ping: 'pong' }.to_json)
     end
   end
 end
