@@ -5,6 +5,6 @@ describe 'Exceptions', js: true, type: :feature do
     visit '/api/raise'
   end
   it 'displays 500 error page' do
-    title.should == 'Action Controller: Exception caught'
+    expect(title).to eq 'Action Controller: Exception caught'
   end
 end
