@@ -8,9 +8,7 @@ require 'rspec/rails'
 RSpec.configure do |config|
   config.mock_with :rspec
   config.expect_with :rspec
-  config.include RSpec::Rails::RequestExampleGroup, type: :request, example_group: {
-    file_path: /spec\/api/
-  }
+  config.include RSpec::Rails::RequestExampleGroup, type: :request, file_path: /spec\/api/
 end
 
 require 'capybara/rspec'
